+++ app/controllers/ctc/portal/submission_pdfs_controller.rb
@@ -9,7 +9,7 @@ class Ctc::Portal::SubmissionPdfsController < Ctc::Portal::BaseAuthenticatedCont
     @document = current_client.documents.find_by(tax_return_id: @submission.tax_return_id, document_type: DocumentTypes::Form1040.key)
     if @submission.present? && !@document.present?
       begin
-        @document = CreateSubmissionPdfJob.perform_now(@submission.id)
+        @document = StateFile::CreateSubmissionPdfJob.perform_now(@submission.id)
       rescue
         error_redirect and return
       end
