+++ app/models/efile_submission.rb
@@ -271,7 +271,7 @@ class EfileSubmission < ApplicationRecord
         age ** 1.25
       end
     retry_wait = backoff + SecureRandom.rand(30)
-    GyrEfiler::SendSubmissionJob.set(wait_until: now + retry_wait).perform_later(self)
+    StateFile::SendSubmissionJob.set(wait_until: now + retry_wait).perform_later(self)
   end
 
   def create_qualifying_dependents
