+++ spec/jobs/state_file/send_reject_resolution_reminder_notification_job_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe SendRejectResolutionReminderNotificationJob, type: :job do
+RSpec.describe StateFile::SendRejectResolutionReminderNotificationJob, type: :job do
   describe "#perform" do
     let(:intake) {
       create :state_file_az_intake,
