+++ spec/tasks/send_reject_resolution_reminder_notifications_spec.rb
@@ -3,7 +3,7 @@ require "rails_helper"
 describe "send_reject_resolution_reminder_notifications:send" do
   let!(:intake) { create :state_file_az_intake, efile_submissions: efile_subimissions }
   before do
-    allow(SendRejectResolutionReminderNotificationJob).to receive(:perform_later).with(intake)
+    allow(StateFile::SendRejectResolutionReminderNotificationJob).to receive(:perform_later).with(intake)
   end
 
   include_context "rake"
@@ -12,7 +12,7 @@ describe "send_reject_resolution_reminder_notifications:send" do
 
     it "enqueues a job" do
       task.invoke
-      expect(SendRejectResolutionReminderNotificationJob).to have_received(:perform_later).with(intake)
+      expect(StateFile::SendRejectResolutionReminderNotificationJob).to have_received(:perform_later).with(intake)
     end
   end
 
@@ -21,7 +21,7 @@ describe "send_reject_resolution_reminder_notifications:send" do
 
     it "doesn't enqueues a job" do
       task.invoke
-      expect(SendRejectResolutionReminderNotificationJob).not_to have_received(:perform_later).with(intake)
+      expect(StateFile::SendRejectResolutionReminderNotificationJob).not_to have_received(:perform_later).with(intake)
     end
   end
 
@@ -30,7 +30,7 @@ describe "send_reject_resolution_reminder_notifications:send" do
 
     it "doesn't enqueues a job" do
       task.invoke
-      expect(SendRejectResolutionReminderNotificationJob).not_to have_received(:perform_later).with(intake)
+      expect(StateFile::SendRejectResolutionReminderNotificationJob).not_to have_received(:perform_later).with(intake)
     end
   end
 end
